Add deletion by shrinking
Now, if the shape's bounding sphere goes below a certain radius, the object will blink and bring up a message indicating that releasing it at this size will cause it to be deleted.

Implementing deletion of objects required the `moddeco` mixin to be adjusted to work around an A-Frame bug; see the comment in `src/js/pseudo-mixins.js` for more information.